kyc/src/questionnaires/service/qnnaire.py
from abc import ABC, abstractmethod
from typing import Any, Dict, Optional, Type

from rest_framework import serializers, status
from rest_framework.response import Response

from ..models import Questionnaire
from ..repo import QuestionnaireRepository

class QuestionnaireService:
    """
    Encapsulates business logic and response construction for questionnaires.
    """
    def __init__(self, repository: QuestionnaireRepository):
        self.repo = repository

    def get_questionnaires(self, params: Dict[str, Any]) -> Response:
        """
        Fetch, serialize, and return questionnaire list.
        """
        pass


    def create_questionnaire(self, data: Dict[str, Any]) -> Response:
        """
        Build, persist, serialize, and return a new questionnaire.
        """
        pass