requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-cacheops==7.1
django-celery-results==2.5.1
django-environ==0.12.0
django-redis==5.4.0
freezegun==1.5.1
funcy==2.0
gunicorn==23.0.0
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pytest==8.3.4
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==4.5.5
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-cacheops==7.1
django-celery-results==2.5.1
django-environ==0.12.0
django-redis==5.4.0
freezegun==1.5.1
funcy==2.0
gunicorn==23.0.0
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pillow==11.3.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pytest==8.3.4
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==4.5.5
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-cacheops==7.1
django-celery-results==2.5.1
django-environ==0.12.0
django-redis==5.4.0
djangorestframework==3.16.0
freezegun==1.5.1
funcy==2.0
gunicorn==23.0.0
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pillow==11.3.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pytest==8.3.4
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==4.5.5
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
